Added logic to mitigate bad string "null" values
If the user hasn't selected a video, we should ideally save a null value indicating that the property doesn't have a value. But Umbraco will convert an actual null value to a string containing the value "null", which then gives a different meaning, and causes a number of Angular errors when interacting with the property editor.