Added Skybrud.Essentials as an explicit dependency
The "Limbo.Umbraco.Video" and "Skybrud.Social.TwentyThree" packages already have a depedency for "Skybrud.Essentials", but it's not the newest version, so we need to add "Skybrud.Essentials" as an explicit dependency to use the newest version. This also improves the support for nullable reference types, which accounts for most of the code changes in this commit.