Implemented new upload link feature
TwentyThree supports creating an upload page in their end, so this feature adds support for showing a "Upload video" link in the property editor, which after selecting one of the configured accounts, will take the user to the upload page. This feature does not allow uploading videos directly from Umbraco.